Bigger text
Apparently there's an optimal number of characters per line.
It's between 60 and 75.
http://www.johndcook.com/blog/2012/09/15/the-paper-is-too-big/

I was getting about 120 with the current style.
Felt like my blog was squished together.
Here are some tweaks that I think make it a bit nicer,
I still feel like it could be better in this area,
but was worried about taking it too far.

But someone else should double check it to make sure it's reasonable.
Especially changing the width,
I don't fully grasp the repercussions of this.