Add a stylesheet for the syntax highlighting.
I played with a lot of them for a while and eventually liked this one best.
It is the normal molokai, but I changed the colour of Generic.Output to
be a lighter, more readable gray, with a tinge of blue that seems to
look nice with its prompt and with our Turing colours.

You can see all the languages and styles by:

```
$ git clone https://github.com/jneen/rouge
$ cd rouge
$ bundle
$ rackup config.ru
$ open 'http://localhost:9292/shell?theme=molokai'
```

Switching "shell" to "ruby" or some other language,
and "theme" to "github" or some other theme.

Get a list of languages by going to the root of that app.
Get a list of themes with:

```
$ ruby -r rouge -e 'puts Rouge::Theme.registry.keys'
thankful_eyes
colorful
base16
base16.dark
base16.solarized
base16.solarized.dark
base16.monokai
base16.monokai.light
github
monokai
molokai
monokai.sublime
```