Fix bug from superclass mismatch on syntax highlighting code
I think I originally was requiring the file before running this,
but then took that out and it still rendered fine,
but I was probably relying on middleman's reloading to do this,
and it probably had that code loaded or cached or something,
so it didn't break until the server was restarted.

Just a guess.